Update README.adoc
The AsciiDoc conditional statements were being interpreted by the AsciiDoc formatter, hiding some of the lines (they were only viewable in raw mode). Adding quotes around the conditionals allows all of the lines of the examples would be displayed. Also added the Important note to be completely clear that the quotes should NOT normally be used.